rename annotator pkg -> refactor, break cli into sub-commands
