+++ waterloo/refactor/__init__.py
@@ -1,0 +1,1 @@
+from .annotations import annotate
